refactor: Separate profile-related controllers (image, info) and update related imports
