+++ src/main/java/oneseoktwojo/ohtalkhae/domain/auth/repository/UserRepository.java
@@ -10,4 +10,6 @@ public interface UserRepository extends JpaRepository<User, Long> {
     Boolean existsByEmail(String email);
     Boolean existsByPhone(String phone);
     User findByUsername(String username);
+    Boolean existsByNickname(String nickname);
+
 }
