+++ src/main/java/oneseoktwojo/ohtalkhae/domain/emoji/controller/EmojiController.java
@@ -1,8 +1,6 @@
 package oneseoktwojo.ohtalkhae.domain.emoji.controller;
 
 import lombok.RequiredArgsConstructor;
-oneseoktwojo.ohtalkhae.domain.emoji.dto.response.EmojiListResponse;
-oneseoktwojo.ohtalkhae.domain.emoji.dto.response.EmojiRegisterResponse;
 import oneseoktwojo.ohtalkhae.domain.emoji.service.EmojiService;
 import oneseoktwojo.ohtalkhae.domain.emoji.dto.request.EmojiRegisterRequest;
 import oneseoktwojo.ohtalkhae.domain.emoji.dto.response.EmojiDetailResponse;
