+++ src/main/java/oneseoktwojo/ohtalkhae/domain/profile/service/ProfileInfoService.java
@@ -1,8 +1,8 @@
 package oneseoktwojo.ohtalkhae.domain.profile.service;
 
 import lombok.RequiredArgsConstructor;
-import oneseoktwojo.ohtalkhae.domain.auth.AuthRepository;
-import oneseoktwojo.ohtalkhae.domain.auth.User;
+import oneseoktwojo.ohtalkhae.domain.auth.entity.User;
+import oneseoktwojo.ohtalkhae.domain.auth.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -11,17 +11,16 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 public class ProfileInfoService {
 
-    private final AuthRepository authRepository;
+    private final UserRepository userRepository;
 
     public void updateNickname(Long userId, String newNickname) {
-
-        if (authRepository.existsByNickname(newNickname)) {
+        if (userRepository.existsByNickname(newNickname)) {
             throw new IllegalArgumentException("이미 사용 중인 닉네임입니다.");
         }
 
-        User user = authRepository.findById(userId)
+        User user = userRepository.findById(userId)
                 .orElseThrow(() -> new IllegalArgumentException("해당 사용자를 찾을 수 없습니다. ID: " + userId));
         user.setNickname(newNickname);
-        authRepository.save(user);
+        userRepository.save(user);
     }
 }
